# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JITAuth Broker",
    "description": "Just-in-time, task-scoped authentication and authorization for AI agents",
    "version": "0.1.0-dev"
  },
  "paths": {
...
    "/tasks": {
      "post": {
        "summary": "Create Task",
        "operationId": "create_task_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "summary": "Get Task",
        "operationId": "get_task_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/classify": {
      "post": {
        "summary": "Classify Task",
        "operationId": "classify_task_tasks__task_id__classify_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/policy-evaluate": {
      "post": {
        "summary": "Evaluate Policy",
        "operationId": "evaluate_policy_tasks__task_id__policy_evaluate_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyDecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/approve": {
      "post": {
        "summary": "Approve Task",
        "operationId": "approve_task_tasks__task_id__approve_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/capabilities": {
      "post": {
        "summary": "Request Capabilities",
        "operationId": "request_capabilities_tasks__task_id__capabilities_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CapabilityResponse"
                  },
                  "title": "Response Request Capabilities Tasks  Task Id  Capabilities Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/execute": {
      "post": {
        "summary": "Execute Tool",
        "operationId": "execute_tool_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/capabilities/{capability_id}/revoke": {
      "post": {
        "summary": "Revoke Capability",
        "operationId": "revoke_capability_capabilities__capability_id__revoke_post",
        "parameters": [
          {
            "name": "capability_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Capability Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/audit": {
      "get": {
        "summary": "Query Audit",
        "operationId": "query_audit_audit_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Id"
            }
          },
          {
            "name": "runtime_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Runtime Id"
            }
          },
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEventResponse"
                  },
                  "title": "Response Query Audit Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApprovalRequest": {
        "properties": {
          "approver_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Approver Id"
          },
          "approved": {
            "type": "boolean",
            "title": "Approved"
          },
          "reduced_scope": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reduced Scope"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "approver_id",
          "approved"
        ],
        "title": "ApprovalRequest"
      },
      "ApprovalResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "approver_id": {
            "type": "string",
            "title": "Approver Id"
          },
          "approved": {
            "type": "boolean",
            "title": "Approved"
          },
          "decided_at": {
            "type": "string",
            "format": "date-time",
            "title": "Decided At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "approver_id",
          "approved",
          "decided_at"
        ],
        "title": "ApprovalResponse"
      },
...
      "CapabilityResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "runtime_id": {
            "type": "string",
            "title": "Runtime Id"
          },
          "target_system": {
            "type": "string",
            "title": "Target System"
          },
          "allowed_actions": {
            "type": "string",
            "title": "Allowed Actions"
          },
          "resource_scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Scope"
          },
          "max_calls": {
            "type": "integer",
            "title": "Max Calls"
          },
          "calls_used": {
            "type": "integer",
            "title": "Calls Used"
          },
          "status": {
            "$ref": "#/components/schemas/CapabilityStatus"
          },
          "issued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Issued At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "runtime_id",
          "target_system",
          "allowed_actions",
          "resource_scope",
          "max_calls",
          "calls_used",
          "status",
          "issued_at",
          "expires_at"
        ],
        "title": "CapabilityResponse"
      },
...
      "ClassifyResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "risk_tier": {
            "$ref": "#/components/schemas/RiskTier"
          },
          "action_classes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Action Classes"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "risk_tier",
          "action_classes"
        ],
        "title": "ClassifyResponse"
      },
      "ExecuteRequest": {
        "properties": {
          "task_id": {
            "type": "string",
            "maxLength": 26,
            "minLength": 1,
            "title": "Task Id"
          },
          "capability_id": {
            "type": "string",
            "maxLength": 26,
            "minLength": 1,
            "title": "Capability Id"
          },
          "tool": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "pattern": "^[\\w.-]+$",
            "title": "Tool"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments"
          },
          "expected_effect": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Effect"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "capability_id",
          "tool"
        ],
        "title": "ExecuteRequest"
      },
      "ExecuteResponse": {
        "properties": {
          "invocation_id": {
            "type": "string",
            "title": "Invocation Id"
          },
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "invocation_id",
          "tool",
          "success"
        ],
        "title": "ExecuteResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PolicyDecisionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "rule_name": {
            "type": "string",
            "title": "Rule Name"
          },
          "effect": {
            "$ref": "#/components/schemas/PolicyEffect"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "evaluated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Evaluated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "rule_name",
          "effect",
          "reason",
          "evaluated_at"
        ],
        "title": "PolicyDecisionResponse"
      },
...
      "RevokeRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason"
          },
          "revoked_by": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Revoked By"
          }
        },
        "type": "object",
        "required": [
          "reason",
          "revoked_by"
        ],
        "title": "RevokeRequest"
      },
      "RevokeResponse": {
        "properties": {
          "capability_id": {
            "type": "string",
            "title": "Capability Id"
          },
          "status": {
            "$ref": "#/components/schemas/CapabilityStatus"
          },
          "revoked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Revoked At"
          }
        },
        "type": "object",
        "required": [
          "capability_id",
          "status",
          "revoked_at"
        ],
        "title": "RevokeResponse"
      },
...
      "TaskCreate": {
        "properties": {
          "requester_type": {
            "type": "string",
            "maxLength": 50,
            "title": "Requester Type",
            "default": "human_user"
          },
          "requester_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Requester Id"
          },
          "requester_auth_context": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Requester Auth Context"
          },
          "runtime_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Runtime Id"
          },
          "runtime_type": {
            "type": "string",
            "maxLength": 100,
            "title": "Runtime Type",
            "default": "llm_orchestrator"
          },
          "runtime_trust_tier": {
            "type": "string",
            "maxLength": 20,
            "title": "Runtime Trust Tier",
            "default": "low"
          },
          "objective": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 1,
            "title": "Objective"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/TaskActionCreate"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Actions"
          },
          "max_actions": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Actions",
            "default": 10
          },
          "time_limit_seconds": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 10.0,
            "title": "Time Limit Seconds",
            "default": 300
          },
          "allow_destructive": {
            "type": "boolean",
            "title": "Allow Destructive",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "requester_id",
          "runtime_id",
          "objective",
          "actions"
        ],
        "title": "TaskCreate"
      },
      "TaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "requester_id": {
            "type": "string",
            "title": "Requester Id"
          },
          "runtime_id": {
            "type": "string",
            "title": "Runtime Id"
          },
          "runtime_type": {
            "type": "string",
            "title": "Runtime Type"
          },
          "objective": {
            "type": "string",
            "title": "Objective"
          },
          "risk_tier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskTier"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/TaskActionResponse"
            },
            "type": "array",
            "title": "Actions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "requester_id",
          "runtime_id",
          "runtime_type",
          "objective",
          "risk_tier",
          "status",
          "actions",
          "created_at",
          "expires_at"
        ],
        "title": "TaskResponse"
      },
...
    }
  }
}